requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-docx>=1.1.0
requests>=2.31.0
Pillow>=10.0.0
pystray>=0.19.4
pytesseract>=0.3.10
pywin32>=306
